+++ lint.sh
@@ -1,0 +1,11 @@
+#!/bin/bash
+gometalinter --vendor --disable-all \
+--enable=vet \
+--enable=vetshadow \
+--enable=golint \
+--enable=ineffassign \
+--enable=goconst \
+--enable=deadcode \
+--enable=gofmt \
+--enable=goimports \
+--tests ./...