+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Changes since v2.2:
 
+- Update CI scripts to separate linting and testing
+  - Now using `gometalinter` for linting
 - Move Go import path from `github.com/bitly/oauth2_proxy` to `github.com/pusher/oauth2_proxy`
 - Repository forked on 27/11/18
   - README updated to include note that this repository is forked